+++ api/httpserver/rest/rest_test.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"fmt"
 	"github.com/spacemeshos/explorer-backend/model"
-	testseed2 "github.com/spacemeshos/explorer-backend/test/testseed"
-	"github.com/spacemeshos/explorer-backend/test/testserver"
 	"os"
 	"testing"
 	"time"
@@ -13,14 +11,18 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 
+	"github.com/spacemeshos/explorer-backend/internal/storage/storagereader"
 	"github.com/spacemeshos/explorer-backend/storage"
+	"github.com/spacemeshos/explorer-backend/test/testseed"
+	"github.com/spacemeshos/explorer-backend/test/testserver"
 )
 
 const testAPIServiceDB = "explorer_test"
 
 var (
 	apiServer *testserver.TestAPIService
-	generator *testseed2.SeedGenerator
+	generator *testseed.SeedGenerator
+	seed      *testseed.TestServerSeed
 	dbPort    = 27017
 )
 
@@ -44,16 +46,22 @@ func TestMain(m *testing.M) {
 		fmt.Println("failed to init storage to mongo", err)
 		os.Exit(1)
 	}
-	seed := testseed2.GetServerSeed()
-	db.OnNetworkInfo(seed.NetID, 2, seed.EpochNumLayers, seed.MaxTransactionPerSecond, seed.LayersDuration, seed.GetPostUnitsSize())
+	seed = testseed.GetServerSeed()
+	db.OnNetworkInfo(seed.NetID, seed.GenesisTime, seed.EpochNumLayers, seed.MaxTransactionPerSecond, seed.LayersDuration, seed.GetPostUnitsSize())
 
-	apiServer, err = testserver.StartTestAPIService(dbPort)
+	dbReader, err := storagereader.NewStorageReader(context.Background(), mongoURL, testAPIServiceDB)
+	if err != nil {
+		fmt.Println("failed to init storage to mongo", err)
+		os.Exit(1)
+	}
+
+	apiServer, err = testserver.StartTestAPIServiceV2(db, dbReader)
+	// old version of app here apiServer, err = testserver.StartTestAPIService(dbPort, db)
 	if err != nil {
 		fmt.Println("failed to start test api service", err)
 		os.Exit(1)
 	}
-	apiServer.Storage = db
-	generator = testseed2.NewSeedGenerator(testseed2.GetServerSeed())
+	generator = testseed.NewSeedGenerator(testseed.GetServerSeed())
 	if err = generator.GenerateEpoches(10); err != nil {
 		fmt.Println("failed to generate epochs", err)
 		os.Exit(1)
